Conditionally prepend Wordfence WAF file in vhost
The variable `wordfence_waf_path` points to Wordfence WAF file which is
used to enable Extended Protection.

Specifying the file in the virtual host configuration has the advantage
that it will only affect a certain site. If the value was specified in
php.ini it could have unwanted effects on other possibly unrelated
sites.

The variable does not have a default value in the role, so the file is
prepended only when it is explicitly defined. Please ensure the file
actually exists or the page will be blank.